Cput

+++ src/afni_setup.c
@@ -177,7 +177,7 @@ STATUS("create initial palettes") ;
 if(PRINT_TRACING)
 { char str[256] ;
   sprintf(str,"found palette label=%s. [len=%d label[0]=%d]",
-          label,strlen(label),(int)label[0]); STATUS(str);
+          label,(int)strlen(label),(int)label[0]); STATUS(str);
   sprintf(str,"nbuf=%d fptr-fbuf=%d",nbuf,fptr-fbuf); STATUS(str);}
 
                ii = label_in_PALTAB( GPT , label ) ; /* an old one? */
